final commit

+++ src/group7_java/school_bussiness_tour_management/views/ManageTeacher.java
@@ -315,14 +315,22 @@ public class ManageTeacher extends javax.swing.JFrame {
                 MessageDialog.showInfoDialog(this, "Vui chọn chọn giáo viên muốn sửa", "Thông báo");
                 return;
             }
-            Teacher selectedTea = TeacherService.getTeacherByIndex(index);
+            String codeTeacher = (String) teacherTable.getValueAt(index, 0);
+            List<Teacher> data_teachers = TeacherService.getAllTeachers();
+            int id = -1;
+            for (Teacher item : data_teachers) {
+                if (item.getCode().equalsIgnoreCase(codeTeacher)) {
+                    id = item.getId();
+                    break;
+                }
+            }
+            Teacher selectedTea = TeacherService.getTeacherById(id);
             String imagePath = selectedTea.getImagePath();
 
             if (!selectedImagePath.equals(imagePath)) {
                 if (!selectedImagePath.trim().equals("")) {
                     selectedTea.setImagePath(selectedImagePath);
                 }
-
             }
 
             String teacherID = this.teacherIDField.getText().trim();
@@ -480,35 +488,46 @@ public class ManageTeacher extends javax.swing.JFrame {
                 MessageDialog.showInfoDialog(this, "Vui chọn chọn giáo viên muốn xóa", "Thông báo");
                 return;
             }
-            Teacher selectedTea = TeacherService.getTeacherByIndex(index);
+            String codeTeacher = (String) teacherTable.getValueAt(index, 0);
+            List<Teacher> data_teachers = TeacherService.getAllTeachers();
+            int id = -1;
+            for (Teacher item : data_teachers) {
+                if (item.getCode().equalsIgnoreCase(codeTeacher)) {
+                    id = item.getId();
+                    break;
+                }
+            }
+            Teacher selectedTea = TeacherService.getTeacherById(id);
             String imagePath = selectedTea.getImagePath();
-
+            String imagePathDefault = "src\\\\group7_java\\\\school_bussiness_tour_management\\\\resources\\\\user.jpg";
             if (imagePath == null) {
                 return;
             }
-            try {
-                // Kiểm tra xem đường dẫn hình ảnh có tồn tại không
-                File imageFile = new File(imagePath);
+            int keyPress = MessageDialog.showConfirmDialog(this, "Bạn có chắc muốn xóa giáo viên " + selectedTea.getLastName() + " " + selectedTea.getFirstName(), "Xác nhận");
+            if (keyPress == 0) {
+                if (!imagePath.equalsIgnoreCase(imagePathDefault)) {
+                    try {
+                        // Kiểm tra xem đường dẫn hình ảnh có tồn tại không
+                        File imageFile = new File(imagePath);
 
-                if (imageFile.exists()) {
-                    // Kiểm tra quyền truy cập và xóa tập tin
-                    if (imageFile.canWrite() && imageFile.canRead()) {
-                        if (!imageFile.delete()) {
+                        if (imageFile.exists()) {
+                            // Kiểm tra quyền truy cập và xóa tập tin
+                            if (imageFile.canWrite() && imageFile.canRead()) {
+                                if (!imageFile.delete()) {
+                                    JOptionPane.showMessageDialog(null, "Ảnh không tồn tại");
+                                }
+                            } else {
+                                JOptionPane.showMessageDialog(null, "Không có quyền truy cập để xóa ảnh.");
+                            }
+                        } else {
                             JOptionPane.showMessageDialog(null, "Ảnh không tồn tại");
                         }
-                    } else {
-                        JOptionPane.showMessageDialog(null, "Không có quyền truy cập để xóa ảnh.");
+                    } catch (SecurityException e) {
+                        e.printStackTrace();
+                        JOptionPane.showMessageDialog(null, "Có lỗi khi xóa ảnh: " + e.getMessage());
                     }
-                } else {
-                    JOptionPane.showMessageDialog(null, "Ảnh không tồn tại");
                 }
-            } catch (SecurityException e) {
-                e.printStackTrace();
-                JOptionPane.showMessageDialog(null, "Có lỗi khi xóa ảnh: " + e.getMessage());
-            }
-            int keyPress = MessageDialog.showConfirmDialog(this, "Bạn có chắc muốn xóa giáo viên " 
-                    + selectedTea.getLastName() + " " + selectedTea.getFirstName(), "Xác nhận");
-            if (keyPress == 0) {
+            
                 TeacherService.deleteTeacher(selectedTea.getId());
                 loadTableData();
                 clearField();
@@ -559,10 +578,8 @@ public class ManageTeacher extends javax.swing.JFrame {
             String teacherCode = (String) teacherTable.getValueAt(index, 0);
             List<Teacher> data_teachers = TeacherService.getAllTeachers();
             int id = -1;
-            for(Teacher item : data_teachers)
-            {
-                if(item.getCode().equalsIgnoreCase(teacherCode))
-                {
+            for (Teacher item : data_teachers) {
+                if (item.getCode().equalsIgnoreCase(teacherCode)) {
                     id = item.getId();
                     break;
                 }
